added new fragment to show film info + some refactoring

+++ app/src/main/java/com/android/course/room/domain/FilmInfo.kt
@@ -6,5 +6,5 @@ data class FilmInfo(
     val year: Int,
     val authors: String,
     val info: String,
-    val rate: Int
+    val rate: Double
 )
